move player plugin to new folder
